LayoutTests/editing/deleting/ios/backspace-last-character-expected.txt

CONSOLE MESSAGE: Activating element and waiting for input session.
CONSOLE MESSAGE: Waiting for element focus.
CONSOLE MESSAGE: Typing 'i'.
CONSOLE MESSAGE: Waiting for first input event.
CONSOLE MESSAGE: Hitting backspace.
CONSOLE MESSAGE: Waiting for second input event.
CONSOLE MESSAGE: Hitting backspace again.
CONSOLE MESSAGE: Waiting for third input event.
CONSOLE MESSAGE: Test complete.
This tests using the software keyboard and pressing 'i' then backspace twice to delete 'Hi'. This should not cause a debug assertion.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inputDidFocus became true
PASS inputEventCount became 1
PASS document.getElementById("input").value is "Hi"
PASS inputEventCount became 2
PASS document.getElementById("input").value is "H"
PASS inputEventCount became 3
PASS document.getElementById("input").value is ""
PASS successfullyParsed is true

TEST COMPLETE
